@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
-Copyright (c) 2015-2020 Rapptz
+Copyright (c) 2015-present Rapptz
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -45,6 +43,7 @@ __all__ = (
     'NotOwner',
     'MessageNotFound',
     'MemberNotFound',
+    'GuildNotFound',
     'UserNotFound',
     'ChannelNotFound',
     'ChannelNotReadable',
@@ -142,7 +141,7 @@ class MissingRequiredArgument(UserInputError):
     """
     def __init__(self, param):
         self.param = param
-        super().__init__('{0.name} is a required argument that is missing.'.format(param))
+        super().__init__(f'{param.name} is a required argument that is missing.')
 
 class TooManyArguments(UserInputError):
     """Exception raised when the command was passed too many arguments and its
@@ -228,7 +227,23 @@ class MemberNotFound(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('Member "{}" not found.'.format(argument))
+        super().__init__(f'Member "{argument}" not found.')
+
+class GuildNotFound(BadArgument):
+    """Exception raised when the guild provided was not found in the bot's cache.
+
+    This inherits from :exc:`BadArgument`
+
+    .. versionadded:: 1.7
+
+    Attributes
+    -----------
+    argument: :class:`str`
+        The guild supplied by the called that was not found
+    """
+    def __init__(self, argument):
+        self.argument = argument
+        super().__init__(f'Guild "{argument}" not found.')
 
 class UserNotFound(BadArgument):
     """Exception raised when the user provided was not found in the bot's
@@ -245,7 +260,7 @@ class UserNotFound(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('User "{}" not found.'.format(argument))
+        super().__init__(f'User "{argument}" not found.')
 
 class MessageNotFound(BadArgument):
     """Exception raised when the message provided was not found in the channel.
@@ -261,7 +276,7 @@ class MessageNotFound(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('Message "{}" not found.'.format(argument))
+        super().__init__(f'Message "{argument}" not found.')
 
 class ChannelNotReadable(BadArgument):
     """Exception raised when the bot does not have permission to read messages
@@ -278,7 +293,7 @@ class ChannelNotReadable(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__("Can't read messages in {}.".format(argument.mention))
+        super().__init__(f"Can't read messages in {argument.mention}.")
 
 class ChannelNotFound(BadArgument):
     """Exception raised when the bot can not find the channel.
@@ -294,7 +309,7 @@ class ChannelNotFound(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('Channel "{}" not found.'.format(argument))
+        super().__init__(f'Channel "{argument}" not found.')
 
 class BadColourArgument(BadArgument):
     """Exception raised when the colour is not valid.
@@ -310,7 +325,7 @@ class BadColourArgument(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('Colour "{}" is invalid.'.format(argument))
+        super().__init__(f'Colour "{argument}" is invalid.')
 
 BadColorArgument = BadColourArgument
 
@@ -328,7 +343,7 @@ class RoleNotFound(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('Role "{}" not found.'.format(argument))
+        super().__init__(f'Role "{argument}" not found.')
 
 class BadInviteArgument(BadArgument):
     """Exception raised when the invite is invalid or expired.
@@ -354,7 +369,7 @@ class EmojiNotFound(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('Emoji "{}" not found.'.format(argument))
+        super().__init__(f'Emoji "{argument}" not found.')
 
 class PartialEmojiConversionFailure(BadArgument):
     """Exception raised when the emoji provided does not match the correct
@@ -371,7 +386,7 @@ class PartialEmojiConversionFailure(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('Couldn\'t convert "{}" to PartialEmoji.'.format(argument))
+        super().__init__(f'Couldn\'t convert "{argument}" to PartialEmoji.')
 
 class BadBoolArgument(BadArgument):
     """Exception raised when a boolean argument was not convertable.
@@ -387,7 +402,7 @@ class BadBoolArgument(BadArgument):
     """
     def __init__(self, argument):
         self.argument = argument
-        super().__init__('{} is not a recognised boolean option'.format(argument))
+        super().__init__(f'{argument} is not a recognised boolean option')
 
 class DisabledCommand(CommandError):
     """Exception raised when the command being invoked is disabled.
@@ -427,7 +442,7 @@ class CommandOnCooldown(CommandError):
     def __init__(self, cooldown, retry_after):
         self.cooldown = cooldown
         self.retry_after = retry_after
-        super().__init__('You are on cooldown. Try again in {:.2f}s'.format(retry_after))
+        super().__init__(f'You are on cooldown. Try again in {retry_after:.2f}s')
 
 class MaxConcurrencyReached(CommandError):
     """Exception raised when the command being invoked has reached its maximum concurrency.
@@ -449,7 +464,7 @@ class MaxConcurrencyReached(CommandError):
         suffix = 'per %s' % name if per.name != 'default' else 'globally'
         plural = '%s times %s' if number > 1 else '%s time %s'
         fmt = plural % (number, suffix)
-        super().__init__('Too many people using this command. It can only be used {} concurrently.'.format(fmt))
+        super().__init__(f'Too many people using this command. It can only be used {fmt} concurrently.')
 
 class MissingRole(CheckFailure):
     """Exception raised when the command invoker lacks a role to run a command.
@@ -466,7 +481,7 @@ class MissingRole(CheckFailure):
     """
     def __init__(self, missing_role):
         self.missing_role = missing_role
-        message = 'Role {0!r} is required to run this command.'.format(missing_role)
+        message = f'Role {missing_role!r} is required to run this command.'
         super().__init__(message)
 
 class BotMissingRole(CheckFailure):
@@ -484,7 +499,7 @@ class BotMissingRole(CheckFailure):
     """
     def __init__(self, missing_role):
         self.missing_role = missing_role
-        message = 'Bot requires the role {0!r} to run this command'.format(missing_role)
+        message = f'Bot requires the role {missing_role!r} to run this command'
         super().__init__(message)
 
 class MissingAnyRole(CheckFailure):
@@ -504,14 +519,14 @@ class MissingAnyRole(CheckFailure):
     def __init__(self, missing_roles):
         self.missing_roles = missing_roles
 
-        missing = ["'{}'".format(role) for role in missing_roles]
+        missing = [f"'{role}'" for role in missing_roles]
 
         if len(missing) > 2:
             fmt = '{}, or {}'.format(", ".join(missing[:-1]), missing[-1])
         else:
             fmt = ' or '.join(missing)
 
-        message = "You are missing at least one of the required roles: {}".format(fmt)
+        message = f"You are missing at least one of the required roles: {fmt}"
         super().__init__(message)
 
 
@@ -533,14 +548,14 @@ class BotMissingAnyRole(CheckFailure):
     def __init__(self, missing_roles):
         self.missing_roles = missing_roles
 
-        missing = ["'{}'".format(role) for role in missing_roles]
+        missing = [f"'{role}'" for role in missing_roles]
 
         if len(missing) > 2:
             fmt = '{}, or {}'.format(", ".join(missing[:-1]), missing[-1])
         else:
             fmt = ' or '.join(missing)
 
-        message = "Bot is missing at least one of the required roles: {}".format(fmt)
+        message = f"Bot is missing at least one of the required roles: {fmt}"
         super().__init__(message)
 
 class NSFWChannelRequired(CheckFailure):
@@ -557,7 +572,7 @@ class NSFWChannelRequired(CheckFailure):
     """
     def __init__(self, channel):
         self.channel = channel
-        super().__init__("Channel '{}' needs to be NSFW for this command to work.".format(channel))
+        super().__init__(f"Channel '{channel}' needs to be NSFW for this command to work.")
 
 class MissingPermissions(CheckFailure):
     """Exception raised when the command invoker lacks permissions to run a
@@ -579,7 +594,7 @@ class MissingPermissions(CheckFailure):
             fmt = '{}, and {}'.format(", ".join(missing[:-1]), missing[-1])
         else:
             fmt = ' and '.join(missing)
-        message = 'You are missing {} permission(s) to run this command.'.format(fmt)
+        message = f'You are missing {fmt} permission(s) to run this command.'
         super().__init__(message, *args)
 
 class BotMissingPermissions(CheckFailure):
@@ -602,7 +617,7 @@ class BotMissingPermissions(CheckFailure):
             fmt = '{}, and {}'.format(", ".join(missing[:-1]), missing[-1])
         else:
             fmt = ' and '.join(missing)
-        message = 'Bot requires {} permission(s) to run this command.'.format(fmt)
+        message = f'Bot requires {fmt} permission(s) to run this command.'
         super().__init__(message, *args)
 
 class BadUnionArgument(UserInputError):
@@ -637,7 +652,7 @@ class BadUnionArgument(UserInputError):
         else:
             fmt = ' or '.join(to_string)
 
-        super().__init__('Could not convert "{0.name}" into {1}.'.format(param, fmt))
+        super().__init__(f'Could not convert "{param.name}" into {fmt}.')
 
 class ArgumentParsingError(UserInputError):
     """An exception raised when the parser fails to parse a user's input.
@@ -661,7 +676,7 @@ class UnexpectedQuoteError(ArgumentParsingError):
     """
     def __init__(self, quote):
         self.quote = quote
-        super().__init__('Unexpected quote mark, {0!r}, in non-quoted string'.format(quote))
+        super().__init__(f'Unexpected quote mark, {quote!r}, in non-quoted string')
 
 class InvalidEndOfQuotedStringError(ArgumentParsingError):
     """An exception raised when a space is expected after the closing quote in a string
@@ -676,7 +691,7 @@ class InvalidEndOfQuotedStringError(ArgumentParsingError):
     """
     def __init__(self, char):
         self.char = char
-        super().__init__('Expected space after closing quotation but received {0!r}'.format(char))
+        super().__init__(f'Expected space after closing quotation but received {char!r}')
 
 class ExpectedClosingQuoteError(ArgumentParsingError):
     """An exception raised when a quote character is expected but not found.
@@ -691,7 +706,7 @@ class ExpectedClosingQuoteError(ArgumentParsingError):
 
     def __init__(self, close_quote):
         self.close_quote = close_quote
-        super().__init__('Expected closing {}.'.format(close_quote))
+        super().__init__(f'Expected closing {close_quote}.')
 
 class ExtensionError(DiscordException):
     """Base exception for extension related errors.
@@ -705,7 +720,7 @@ class ExtensionError(DiscordException):
     """
     def __init__(self, message=None, *args, name):
         self.name = name
-        message = message or 'Extension {!r} had an error.'.format(name)
+        message = message or f'Extension {name!r} had an error.'
         # clean-up @everyone and @here mentions
         m = message.replace('@everyone', '@\u200beveryone').replace('@here', '@\u200bhere')
         super().__init__(m, *args)
@@ -716,7 +731,7 @@ class ExtensionAlreadyLoaded(ExtensionError):
     This inherits from :exc:`ExtensionError`
     """
     def __init__(self, name):
-        super().__init__('Extension {!r} is already loaded.'.format(name), name=name)
+        super().__init__(f'Extension {name!r} is already loaded.', name=name)
 
 class ExtensionNotLoaded(ExtensionError):
     """An exception raised when an extension was not loaded.
@@ -724,7 +739,7 @@ class ExtensionNotLoaded(ExtensionError):
     This inherits from :exc:`ExtensionError`
     """
     def __init__(self, name):
-        super().__init__('Extension {!r} has not been loaded.'.format(name), name=name)
+        super().__init__(f'Extension {name!r} has not been loaded.', name=name)
 
 class NoEntryPointError(ExtensionError):
     """An exception raised when an extension does not have a ``setup`` entry point function.
@@ -732,7 +747,7 @@ class NoEntryPointError(ExtensionError):
     This inherits from :exc:`ExtensionError`
     """
     def __init__(self, name):
-        super().__init__("Extension {!r} has no 'setup' function.".format(name), name=name)
+        super().__init__(f"Extension {name!r} has no 'setup' function.", name=name)
 
 class ExtensionFailed(ExtensionError):
     """An exception raised when an extension failed to load during execution of the module or ``setup`` entry point.
@@ -791,4 +806,4 @@ class CommandRegistrationError(ClientException):
         self.name = name
         self.alias_conflict = alias_conflict
         type_ = 'alias' if alias_conflict else 'command'
-        super().__init__('The {} {} is already an existing command or alias.'.format(type_, name))
+        super().__init__(f'The {type_} {name} is already an existing command or alias.')
