@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
-Copyright (c) 2015-2020 Rapptz
+Copyright (c) 2015-present Rapptz
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -26,6 +24,10 @@ DEALINGS IN THE SOFTWARE.
 
 from . import utils
 from .mixins import Hashable
+
+__all__ = (
+    'Object',
+)
 
 class Object(Hashable):
     """Represents a generic Discord object.
@@ -65,12 +67,12 @@ class Object(Hashable):
         try:
             id = int(id)
         except ValueError:
-            raise TypeError('id parameter must be convertable to int not {0.__class__!r}'.format(id)) from None
+            raise TypeError(f'id parameter must be convertable to int not {id.__class__!r}') from None
         else:
             self.id = id
 
     def __repr__(self):
-        return '<Object id=%r>' % self.id
+        return f'<Object id={self.id!r}>'
 
     @property
     def created_at(self):
