@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
-Copyright (c) 2015-2020 Rapptz
+Copyright (c) 2015-present Rapptz
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -44,10 +42,6 @@ class _PartialTemplateState:
         self.http = _FriendlyHttpAttributeErrorHelper()
 
     @property
-    def is_bot(self):
-        return self.__state.is_bot
-
-    @property
     def shard_count(self):
         return self.__state.shard_count
 
@@ -76,7 +70,7 @@ class _PartialTemplateState:
         return []
 
     def __getattr__(self, attr):
-        raise AttributeError('PartialTemplateState does not support {0!r}.'.format(attr))
+        raise AttributeError(f'PartialTemplateState does not support {attr!r}.')
 
 class Template:
     """Represents a Discord template.
@@ -96,16 +90,19 @@ class Template:
     creator: :class:`User`
         The creator of the template.
     created_at: :class:`datetime.datetime`
-        When the template was created.
+        An aware datetime in UTC representing when the template was created.
     updated_at: :class:`datetime.datetime`
-        When the template was last updated (referred to as "last synced" in the client).
+        An aware datetime in UTC representing when the template was last updated.
+        This is referred to as "last synced" in the official Discord client.
     source_guild: :class:`Guild`
         The source guild.
     """
 
     def __init__(self, *, state, data):
         self._state = state
+        self._store(data)
 
+    def _store(self, data):
         self.code = data['code']
         self.uses = data['usage_count']
         self.name =  data['name']
@@ -117,11 +114,16 @@ class Template:
         self.updated_at = parse_time(data.get('updated_at'))
 
         id = _get_as_snowflake(data, 'source_guild_id')
-        source_serialised = data['serialized_source_guild']
-        source_serialised['id'] = id
-        state = _PartialTemplateState(state=self._state)
 
-        self.source_guild = Guild(data=source_serialised, state=state)
+        guild = self._state._get_guild(id)
+
+        if guild is None:
+            source_serialised = data['serialized_source_guild']
+            source_serialised['id'] = id
+            state = _PartialTemplateState(state=self._state)
+            guild = Guild(data=source_serialised, state=state)
+
+        self.source_guild = guild
 
     def __repr__(self):
         return '<Template code={0.code!r} uses={0.uses} name={0.name!r}' \
@@ -147,9 +149,9 @@ class Template:
 
         Raises
         ------
-        :exc:`.HTTPException`
+        HTTPException
             Guild creation failed.
-        :exc:`.InvalidArgument`
+        InvalidArgument
             Invalid icon image format given. Must be PNG or JPG.
 
         Returns
@@ -161,10 +163,81 @@ class Template:
         if icon is not None:
             icon = _bytes_to_base64_data(icon)
 
-        if region is None:
-            region = VoiceRegion.us_west.value
-        else:
-            region = region.value
+        region = region or VoiceRegion.us_west
+        region_value = region.value
 
-        data = await self._state.http.create_from_template(self.code, name, region, icon)
+        data = await self._state.http.create_from_template(self.code, name, region_value, icon)
         return Guild(data=data, state=self._state)
+
+    async def sync(self):
+        """|coro|
+
+        Sync the template to the guild's current state.
+
+        You must have the :attr:`~Permissions.manage_guild` permission in the
+        source guild to do this.
+
+        .. versionadded:: 1.7
+
+        Raises
+        -------
+        HTTPException
+            Editing the template failed.
+        Forbidden
+            You don't have permissions to edit the template.
+        NotFound
+            This template does not exist.
+        """
+
+        data = await self._state.http.sync_template(self.source_guild.id, self.code)
+        self._store(data)
+
+    async def edit(self, **kwargs):
+        """|coro|
+
+        Edit the template metadata.
+
+        You must have the :attr:`~Permissions.manage_guild` permission in the
+        source guild to do this.
+
+        .. versionadded:: 1.7
+
+        Parameters
+        ------------
+        name: Optional[:class:`str`]
+            The template's new name.
+        description: Optional[:class:`str`]
+            The template's description.
+
+        Raises
+        -------
+        HTTPException
+            Editing the template failed.
+        Forbidden
+            You don't have permissions to edit the template.
+        NotFound
+            This template does not exist.
+        """
+        data = await self._state.http.edit_template(self.source_guild.id, self.code, kwargs)
+        self._store(data)
+
+    async def delete(self):
+        """|coro|
+
+        Delete the template.
+
+        You must have the :attr:`~Permissions.manage_guild` permission in the
+        source guild to do this.
+
+        .. versionadded:: 1.7
+
+        Raises
+        -------
+        HTTPException
+            Editing the template failed.
+        Forbidden
+            You don't have permissions to edit the template.
+        NotFound
+            This template does not exist.
+        """
+        await self._state.http.delete_template(self.source_guild.id, self.code)
