@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
-Copyright (c) 2015-2020 Rapptz
+Copyright (c) 2015-present Rapptz
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -25,6 +23,10 @@ DEALINGS IN THE SOFTWARE.
 """
 
 import asyncio
+
+__all__ = (
+    'Typing',
+)
 
 def _typing_done_callback(fut):
     # just retrieve any exception and call it a day
