@@ -1,17 +1,15 @@
-# -*- coding: utf-8 -*-
-
 """
 discord.ext.commands
 ~~~~~~~~~~~~~~~~~~~~~
 
 An extension module to facilitate creation of bot commands.
 
-:copyright: (c) 2015-2020 Rapptz
+:copyright: (c) 2015-present Rapptz
 :license: MIT, see LICENSE for more details.
 """
 
-from .bot import Bot, AutoShardedBot, when_mentioned, when_mentioned_or
-from .context import Context
+from .bot import *
+from .context import *
 from .core import *
 from .errors import *
 from .help import *
