@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
-Copyright (c) 2015-2020 Rapptz
+Copyright (c) 2015-present Rapptz
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -25,6 +23,10 @@ DEALINGS IN THE SOFTWARE.
 """
 
 from .iterators import ReactionIterator
+
+__all__ = (
+    'Reaction',
+)
 
 class Reaction:
     """Represents a reaction to a message.
