@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
-Copyright (c) 2015-2020 Rapptz
+Copyright (c) 2015-present Rapptz
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -45,6 +43,11 @@ from .errors import (
 
 from . import utils
 from .enums import Status
+
+__all__ = (
+    'AutoShardedClient',
+    'ShardInfo',
+)
 
 log = logging.getLogger(__name__)
 
@@ -319,7 +322,7 @@ class AutoShardedClient(Client):
 
     def _get_state(self, **options):
         return AutoShardedConnectionState(dispatch=self.dispatch,
-                                          handlers=self._handlers, syncer=self._syncer,
+                                          handlers=self._handlers,
                                           hooks=self._hooks, http=self.http, loop=self.loop, **options)
 
     @property
@@ -413,7 +416,7 @@ class AutoShardedClient(Client):
 
         self._connection.shard_count = self.shard_count
 
-        shard_ids = self.shard_ids if self.shard_ids else range(self.shard_count)
+        shard_ids = self.shard_ids or range(self.shard_count)
         self._connection.shard_ids = shard_ids
 
         for shard_id in shard_ids:
